moved map to it's own package
This is necessary to split this class in smaller parts
