+++ app/src/main/java/ch/epfl/sdp/TabActivity.java
@@ -11,7 +11,8 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentPagerAdapter;
-import androidx.viewpager.widget.ViewPager;
+
+import ch.epfl.sdp.Map.MapFragment;
 
 public class TabActivity extends AppCompatActivity {
